tools/cg-schema/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cg-schema"
version = "0.1.0"
description = "ClickGraph Schema Designer - ML-powered schema discovery tool"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "ClickGraph Team", email = "team@clickgraph.dev"}
]
keywords = ["clickgraph", "schema", "graph", "ner", "gliner"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "clickhouse-connect>=0.6.0",
    "gliner>=0.2.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "rich>=13.0.0",
]

[project.scripts]
cg-schema = "cg_schema.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cg_schema*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
